+++ DSCResources/MSFT_xSQLServerNetwork/MSFT_xSQLServerNetwork.psm1
@@ -1,6 +1,11 @@
 Import-Module -Name (Join-Path -Path (Split-Path (Split-Path $PSScriptRoot -Parent) -Parent) `
-        -ChildPath 'xSQLServerHelper.psm1') `
-    -Force
+        -ChildPath 'xSQLServerHelper.psm1') -Force
+Import-Module -Name (Join-Path -Path (Split-Path -Path $PSScriptRoot -Parent) `
+        -ChildPath 'CommonResourceHelper.psm1')
+
+# Load localized string data
+$script:localizedData = Get-LocalizedData -ResourceName 'MSFT_xSQLServerNetwork'
+
 <#
     .SYNOPSIS
     Returns the current state of the SQL Server network properties.
@@ -11,7 +16,7 @@ Import-Module -Name (Join-Path -Path (Split-Path (Split-Path $PSScriptRoot -Pare
     .PARAMETER ProtocolName
     The name of network protocol to be configured. Only tcp is currently supported.
 #>
-Function Get-TargetResource
+function Get-TargetResource
 {
     [CmdletBinding()]
     [OutputType([System.Collections.Hashtable])]
@@ -34,20 +39,20 @@ Function Get-TargetResource
 
         $managedComputerObject = New-Object -TypeName Microsoft.SqlServer.Management.Smo.Wmi.ManagedComputer
 
-        Write-Verbose "Getting network protocol [$ProtocolName] for SQL instance [$InstanceName]."
+        Write-Verbose -Message ($script:localizedData.GetNetworkProtocol -f $ProtocolName, $InstanceName)
         $tcp = $managedComputerObject.ServerInstances[$InstanceName].ServerProtocols[$ProtocolName]
 
-        Write-Verbose "Reading current network properties."
+        Write-Verbose -Message $script:localizedData.ReadingNetworkProperties
         $returnValue = @{
-            InstanceName    = $InstanceName
+            InstanceName = $InstanceName
             ProtocolName    = $ProtocolName
             IsEnabled       = $tcp.IsEnabled
-            TcpDynamicPorts = $tcp.IPAddresses['IPAll'].IPAddressProperties['TcpDynamicPorts'].Value
+            TcpDynamicPort  = ($tcp.IPAddresses['IPAll'].IPAddressProperties['TcpDynamicPorts'].Value -ge 0)
             TcpPort         = $tcp.IPAddresses['IPAll'].IPAddressProperties['TcpPort'].Value
         }
 
         $returnValue.Keys | ForEach-Object {
-            Write-Verbose "$_ = $($returnValue[$_])"
+            Write-Verbose -Message "$_ = $($returnValue[$_])"
         }
     }
     finally
@@ -74,16 +79,15 @@ Function Get-TargetResource
     .PARAMETER IsEnabled
     Enables or disables the network protocol.
 
-    .PARAMETER TcpDynamicPorts
-    Set the value to '0' if dynamic ports should be used.
-    If static port should be used set this to a empty string value.
-    Value can not be set to '0' if TcpPort is also set to a value.
+    .PARAMETER TcpDynamicPort
+    Specifies whether the SQL Server instance should use a dynamic port.
+    Value cannot be set to $true if TcpPort is set to a non-empty string.
 
     .PARAMETER TcpPort
     The TCP port(s) that SQL Server should be listening on.
     If the IP address should listen on more than one port, list all ports
     separated with a comma ('1433,1500,1501'). To use this parameter set
-    TcpDynamicPorts to the value '' (empty string).
+    TcpDynamicPort to 'False'.
 
     .PARAMETER RestartService
     If set to $true then SQL Server and dependent services will be restarted
@@ -93,7 +97,7 @@ Function Get-TargetResource
     Timeout value for restarting the SQL Server services. The default value
     is 120 seconds.
 #>
-Function Set-TargetResource
+function Set-TargetResource
 {
     [CmdletBinding()]
     param
@@ -117,9 +121,8 @@ Function Set-TargetResource
         $IsEnabled,
 
         [Parameter()]
-        [ValidateSet('0', '')]
-        [System.String]
-        $TcpDynamicPorts,
+        [System.Boolean]
+        $TcpDynamicPort,
 
         [Parameter()]
         [System.String]
@@ -134,9 +137,10 @@ Function Set-TargetResource
         $RestartTimeout = 120
     )
 
-    if ($TcpDynamicPorts -eq '0' -and $TcpPort)
+    if ($TcpDynamicPort -and $TcpPort)
     {
-        throw New-TerminatingError -ErrorType UnableToUseBothDynamicAndStaticPort -ErrorCategory InvalidOperation
+        $errorMessage = $script:localizedData.ErrorDynamicAndStaticPortSpecified
+        New-InvalidOperationException -Message $errorMessage
     }
 
     $getTargetResourceResult = Get-TargetResource -InstanceName $InstanceName -ProtocolName $ProtocolName
@@ -146,10 +150,10 @@ Function Set-TargetResource
         $applicationDomainObject = Register-SqlWmiManagement -SQLInstanceName $InstanceName
 
         $desiredState = @{
-            InstanceName    = $InstanceName
+            InstanceName = $InstanceName
             ProtocolName    = $ProtocolName
             IsEnabled       = $IsEnabled
-            TcpDynamicPorts = $TcpDynamicPorts
+            TcpDynamicPort  = $TcpDynamicPort
             TcpPort         = $TcpPort
         }
 
@@ -157,42 +161,45 @@ Function Set-TargetResource
 
         $managedComputerObject = New-Object Microsoft.SqlServer.Management.Smo.Wmi.ManagedComputer
 
-        Write-Verbose "Getting [$ProtocolName] network protocol for [$InstanceName] SQL instance"
+        Write-Verbose -Message ($script:localizedData.GetNetworkProtocol -f $ProtocolName, $InstanceName)
         $tcp = $managedComputerObject.ServerInstances[$InstanceName].ServerProtocols[$ProtocolName]
 
-        Write-Verbose 'Checking [IsEnabled] property.'
+        Write-Verbose -Message ($script:localizedData.CheckingProperty -f 'IsEnabled')
         if ($desiredState.IsEnabled -ine $getTargetResourceResult.IsEnabled)
         {
-            Write-Verbose "Updating [IsEnabled] from $($getTargetResourceResult.IsEnabled) to $($desiredState.IsEnabled)."
+            Write-Verbose -Message ($script:localizedData.UpdatingProperty -f 'IsEnabled', $getTargetResourceResult.IsEnabled, $desiredState.IsEnabled)
             $tcp.IsEnabled = $desiredState.IsEnabled
             $tcp.Alter()
 
             $isRestartNeeded = $true
         }
 
-        Write-Verbose 'Checking [TcpDynamicPorts] property.'
-        if ($desiredState.TcpDynamicPorts -ine $getTargetResourceResult.TcpDynamicPorts)
+        Write-Verbose -Message ($script:localizedData.CheckingProperty -f 'TcpDynamicPort')
+        if ($desiredState.TcpDynamicPort -ne $getTargetResourceResult.TcpDynamicPort)
         {
-            $fromTcpDynamicPortsValue = $getTargetResourceResult.TcpDynamicPorts
-            if ($fromTcpDynamicPortsValue -eq '')
-            {
-                $fromTcpDynamicPortsValue = 'none'
+            # Translates the current and desired state to a string for display
+            $dynamicPortDisplayValueTable = @{
+                $true = 'enabled'
+                $false = 'disabled'
             }
 
-            $toTcpDynamicPortsValue = $desiredState.TcpDynamicPorts
-            if ($toTcpDynamicPortsValue -eq '')
-            {
-                $toTcpDynamicPortsValue = 'none'
+            # Translates the desired state to a valid value
+            $desiredDynamicPortValue = @{
+                $true = '0'
+                $false = ''
             }
 
-            Write-Verbose "Updating [TcpDynamicPorts] from $($fromTcpDynamicPortsValue) to $($toTcpDynamicPortsValue)."
-            $tcp.IPAddresses['IPAll'].IPAddressProperties['TcpDynamicPorts'].Value = $desiredState.TcpDynamicPorts
+            $fromTcpDynamicPortDisplayValue = $dynamicPortDisplayValueTable[$getTargetResourceResult.TcpDynamicPort]
+            $toTcpDynamicPortDisplayValue = $dynamicPortDisplayValueTable[$desiredState.TcpDynamicPort]
+
+            Write-Verbose -Message ($script:localizedData.UpdatingProperty -f 'TcpDynamicPorts', $fromTcpDynamicPortDisplayValue, $toTcpDynamicPortDisplayValue)
+            $tcp.IPAddresses['IPAll'].IPAddressProperties['TcpDynamicPorts'].Value = $desiredDynamicPortValue[$desiredState.TcpDynamicPort]
             $tcp.Alter()
 
             $isRestartNeeded = $true
         }
 
-        Write-Verbose 'Checking [TcpPort property].'
+        Write-Verbose -Message ($script:localizedData.CheckingProperty -f 'TcpPort')
         if ($desiredState.TcpPort -ine $getTargetResourceResult.TcpPort)
         {
             $fromTcpPort = $getTargetResourceResult.TcpPort
@@ -207,7 +214,7 @@ Function Set-TargetResource
                 $toTcpPort = 'none'
             }
 
-            Write-Verbose "Updating [TcpPort] from $($fromTcpPort) to $($toTcpPort)."
+            Write-Verbose -Message ($script:localizedData.UpdatingProperty -f 'TcpPort', $fromTcpPort, $toTcpPort)
             $tcp.IPAddresses['IPAll'].IPAddressProperties['TcpPort'].Value = $desiredState.TcpPort
             $tcp.Alter()
 
@@ -243,16 +250,15 @@ Function Set-TargetResource
     .PARAMETER IsEnabled
     Enables or disables the network protocol.
 
-    .PARAMETER TcpDynamicPorts
-    Set the value to '0' if dynamic ports should be used.
-    If static port should be used set this to a empty string value.
-    Value can not be set to '0' if TcpPort is also set to a value.
+    .PARAMETER TcpDynamicPort
+    Specifies whether the SQL Server instance should use a dynamic port.
+    Value cannot be set to $true if TcpPort is set to a non-empty string.
 
     .PARAMETER TcpPort
     The TCP port(s) that SQL Server should be listening on.
     If the IP address should listen on more than one port, list all ports
     separated with a comma ('1433,1500,1501'). To use this parameter set
-    TcpDynamicPorts to the value '' (empty string).
+    TcpDynamicPort to 'False'.
 
     .PARAMETER RestartService
     If set to $true then SQL Server and dependent services will be restarted
@@ -266,7 +272,7 @@ Function Set-TargetResource
 
     Not used in Test-TargetResource.
 #>
-Function Test-TargetResource
+function Test-TargetResource
 {
     [CmdletBinding()]
     [OutputType([System.Boolean])]
@@ -290,9 +296,8 @@ Function Test-TargetResource
         $IsEnabled,
 
         [Parameter()]
-        [ValidateSet('0', '')]
-        [System.String]
-        $TcpDynamicPorts,
+        [System.Boolean]
+        $TcpDynamicPort,
 
         [Parameter()]
         [System.String]
@@ -307,20 +312,21 @@ Function Test-TargetResource
         $RestartTimeout = 120
     )
 
-    if ($TcpDynamicPorts -eq '0' -and $TcpPort)
+    if ($TcpDynamicPort -and $TcpPort)
     {
-        throw New-TerminatingError -ErrorType UnableToUseBothDynamicAndStaticPort -ErrorCategory InvalidOperation
+        $errorMessage = $script:localizedData.ErrorDynamicAndStaticPortSpecified
+        New-InvalidOperationException -Message $errorMessage
     }
 
     $getTargetResourceResult = Get-TargetResource -InstanceName $InstanceName -ProtocolName $ProtocolName
 
-    Write-Verbose "Comparing desired state with current state."
+    Write-Verbose -Message $script:localizedData.CompareStates
 
     $isInDesiredState = $true
 
     if ($ProtocolName -ne $getTargetResourceResult.ProtocolName)
     {
-        Write-Verbose ('Expected protocol to be ''{0}'', but was ''{1}''.' -f $ProtocolName, $getTargetResourceResult.ProtocolName)
+        Write-Verbose -Message ($script:localizedData.ExpectedPropertyValue -f 'ProtocolName', $ProtocolName, $getTargetResourceResult.ProtocolName)
 
         $isInDesiredState = $false
     }
@@ -334,17 +340,17 @@ Function Test-TargetResource
                 $false = 'disabled'
             }
 
-            Write-Verbose ('Expected protocol to be {0}, but was {1}.' -f $evaluateEnableOrDisable[$IsEnabled], $evaluateEnableOrDisable[$getTargetResourceResult.IsEnabled])
+            Write-Verbose -Message ($script:localizedData.ExpectedPropertyValue -f 'IsEnabled', $evaluateEnableOrDisable[$IsEnabled], $evaluateEnableOrDisable[$getTargetResourceResult.IsEnabled])
 
             $isInDesiredState = $false
         }
     }
 
-    if ($PSBoundParameters.ContainsKey('TcpDynamicPorts'))
+    if ($PSBoundParameters.ContainsKey('TcpDynamicPort'))
     {
-        if ($TcpDynamicPorts -eq '0' -and $getTargetResourceResult.TcpDynamicPorts -eq '')
+        if ($TcpDynamicPort -and $getTargetResourceResult.TcpDynamicPort -eq $false)
         {
-            Write-Verbose 'Expected tcp dynamic port to be used, but it was not.'
+            Write-Verbose -Message ($script:localizedData.ExpectedPropertyValue -f 'TcpDynamicPort', $TcpDynamicPort, $getTargetResourceResult.TcpDynamicPort)
 
             $isInDesiredState = $false
         }
@@ -354,13 +360,13 @@ Function Test-TargetResource
     {
         if ($getTargetResourceResult.TcpPort -eq '')
         {
-            Write-Verbose 'Expected tcp static port to be used, but it was not.'
+            Write-Verbose -Message ($script:localizedData.ExpectedPropertyValue -f 'TcpPort', $TcpPort, $getTargetResourceResult.TcpPort)
 
             $isInDesiredState = $false
         }
         elseif ($TcpPort -ne $getTargetResourceResult.TcpPort)
         {
-            Write-Verbose ('Expected tcp static port to be ''{0}'', but was ''{1}''.' -f $TcpPort, $getTargetResourceResult.TcpPort)
+            Write-Verbose -Message ($script:localizedData.ExpectedPropertyValue -f 'TcpPort', $TcpPort, $getTargetResourceResult.TcpPort)
 
             $isInDesiredState = $false
         }
@@ -368,7 +374,7 @@ Function Test-TargetResource
 
     if ($isInDesiredState)
     {
-        Write-Verbose "In desired state."
+        Write-Verbose -Message ($script:localizedData.InDesiredState)
     }
 
     return $isInDesiredState