+++ CHANGELOG.md
@@ -20,6 +20,14 @@
 - Added the CommonTestHelper.psm1 to store common testing functions.
   - Added the Import-SQLModuleStub function to ensure the correct version of the
     module stubs are loaded ([issue #784](https://github.com/PowerShell/xSQLServer/issues/784)).
+- Changes to xSQLServerNetwork
+  - BREAKING CHANGE: Renamed parameter TcpDynamicPorts to TcpDynamicPort and
+    changed type to Boolean.
+  - Resolved issue when switching from dynamic to static port.
+    configuration ([issue #534](https://github.com/PowerShell/xSQLServer/issues/534)).
+  - Added localization (en-US) for all strings in resource and unit tests
+    ([issue #618](https://github.com/PowerShell/xSQLServer/issues/618)).
+  - Updated examples to reflect new parameters.
 
 ## 8.2.0.0